# api/docs/swagger.yaml
definitions:
...
  model.CreateQuestionReq:
    properties:
      answer:
        type: string
      answer_image_url:
        type: string
      option_image_url:
        $ref: '#/definitions/model.Option'
      options:
        $ref: '#/definitions/model.Option'
      question_image_url:
        type: string
      question_text:
        type: string
      subject_id:
        type: string
      type:
        type: string
    type: object
...
  model.Error:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  model.GetQuestionsResp:
    properties:
      questions:
        items:
          $ref: '#/definitions/model.Question'
        type: array
    type: object
...
  model.Incorrect:
    properties:
      name:
        type: string
      nomer:
        type: integer
    type: object
  model.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.Option:
    properties:
      a:
        type: string
      b:
        type: string
      c:
        type: string
      d:
        type: string
    type: object
  model.Question:
    properties:
      answer:
        type: string
      answer_image_url:
        type: string
      id:
        type: string
      option_image_url:
        $ref: '#/definitions/model.Option'
      options:
        $ref: '#/definitions/model.Option'
      question_image_url:
        type: string
      question_text:
        type: string
      subject_id:
        type: string
      type:
        type: string
    type: object
...
  model.QuestionsStatus:
    properties:
      correct:
        type: integer
      incorrect:
        type: integer
      incorrect_questions:
        items:
          $ref: '#/definitions/model.Incorrect'
        type: array
    type: object
  model.RegisterReq:
    properties:
      lastname:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  model.Status:
    properties:
      message:
        type: string
    type: object
...
  model.Token:
    properties:
      access_token:
        type: string
    type: object
  model.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user_id:
        type: string
    type: object
  model.UpdateQuestionReq:
    properties:
      answer:
        type: string
      answer_image_url:
        type: string
      option)image_url:
        $ref: '#/definitions/model.Option'
      options:
        $ref: '#/definitions/model.Option'
      question_image_url:
        type: string
      question_text:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access/refresh tokens
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Tokens'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/model.Error'
        "404":
          description: User not found or invalid credentials
          schema:
            $ref: '#/definitions/model.Error'
      summary: User login
      tags:
      - auth
  /questions:
    get:
      description: Berilgan ID yoki fan ID bo‘yicha savollarni olish
      parameters:
      - description: Savol IDsi
        in: query
        name: id
        type: string
      - description: Fan IDsi
        in: query
        name: subject_id
        type: string
      - description: Savol turi
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetQuestionsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Savollarni olish
      tags:
      - Questions
  /questions/create:
    post:
      consumes:
      - application/json
      description: Yangi savol qo'shish uchun API
      parameters:
      - description: Savol yaratish ma'lumotlari
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateQuestionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Status'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Yangi savol yaratish
      tags:
      - Questions
  /questions/delete/{id}:
    delete:
      description: Berilgan ID bo‘yicha savolni o‘chirish
      parameters:
      - description: Savol IDsi
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Status'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Savolni o‘chirish
      tags:
      - Questions
  /questions/image/upload:
    post:
      consumes:
      - multipart/form-data
      description: Rasmni serverga yuklab, Minio'ga saqlaydi va URL qaytaradi
      parameters:
      - description: Yuklanadigan rasm fayli
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Yuklangan rasm URL'si
          schema:
            type: string
        "400":
          description: Faylni yuklab olishda xatolik
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Faylni yuklash
      tags:
      - Files
  /questions/update/{id}:
    put:
      consumes:
      - application/json
      description: Berilgan ID bo‘yicha savolni yangilash
      parameters:
      - description: Savol IDsi
        in: path
        name: id
        required: true
        type: string
      - description: Savolni yangilash ma'lumotlari
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateQuestionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Status'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Savolni yangilash
      tags:
      - Questions
  /questions/upload:
    post:
      consumes:
      - multipart/form-data
      description: Foydalanuvchi Excel faylini yuklab, uni serverda qayta ishlash
      parameters:
      - description: Excel fayl
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.QuestionsStatus'
        "400":
          description: Faylni yuklashda yoki saqlashda xatolik
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Faylni yuklash va tahlil qilish
      tags:
      - Files
  /refresh-token:
    get:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Refresh Token
        in: query
        name: refresh_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Token'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/model.Error'
      summary: Refresh access token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with necessary credentials
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Tokens'
        "400":
          description: 'Bad request: invalid input'
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.Error'
      summary: User registration
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Tokenni faqat o‘zi yozing, masalan: eyJhbGciOiJIUzI1NiIs..'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"